Cast float to int explicitly.

+++ itermocil.py
@@ -252,7 +252,7 @@ end hack_sid"""
         elif layout == 'tiled':
 
             vertical_splits = int(ceil((num_panes / 2.0))) - 1
-            second_columns = num_panes / 2
+            second_columns = int(num_panes / 2)
 
             for p in range(0, vertical_splits):
                 pp = (p * 2) + 1
